Add test cases for mixed data
